+++ point3/point3.cpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-point3_t::point3_t(const float &x, const float &y, const float &z)
+point3_t::point3_t(const double &x, const double &y, const double &z)
 	: coords{ x, y, z }
 {}
 
@@ -14,12 +14,12 @@ point3_t::point3_t(const glm::vec3 &coords)
 	: coords{ coords }
 {}
 
-const float point3_t::operator[](const size_t &i) const
+const double point3_t::operator[](const size_t &i) const
 {
 	return coords[i];
 }
 
-point3_t::point3_t(const polar_t &p, const float &r)
+point3_t::point3_t(const polar_t &p, const double &r)
 	: coords{
 		r * DCOS(p[0]) * DSIN(p[1]),
 		r * DSIN(p[0]) * DSIN(p[1]),
